menuitem components done
