+++ app/src/main/java/reservant_mobile/ui/components/EmployeeComponents.kt
@@ -46,7 +46,7 @@ fun EmployeeCard(
 
     when {
         showConfirmDeletePopup -> {
-            CountDownPopup(
+            DeleteCountdownPopup(
                 icon = Icons.Filled.DeleteForever,
                 title = stringResource(id = R.string.confirm_delete_title),
                 text = stringResource(id = R.string.confirm_delete_text),
